This is my capstone project I named Cartrack
It allows the user to login as either a salesperson of inventory management employee.  Salespeople can track customers and search the database for vehicles that match what their customers are looking for or look up customer information. Inventory management employees can add or remove vehicles from the database as they are brought in or sold. When a vehicle is brought in that matches a vehicle that a customer is looking for Cartrack will send an email to the salesperson who set up that customer informing then that a similar vehicle has just been added to inventory.